requirements.txt
python-telegram-bot==20.7
google-genai>=0.3.0,<1.0.0  # Migrated from google-generativeai
python-dotenv==1.0.0
aiohttp==3.9.0